+++ forui/test/src/widgets/button/button_golden_test.dart
@@ -23,8 +23,8 @@ void main() {
                 child: FButton(
                   label: const Text('Button'),
                   style: variant,
-                  prefix: FButtonIcon(icon: FAssets.icons.circlePlay),
-                  suffix: FButtonIcon(icon: FAssets.icons.circleStop),
+                  prefix: FIcon(FAssets.icons.circlePlay),
+                  suffix: FIcon(FAssets.icons.circleStop),
                   onPress: () {},
                 ),
               ),
@@ -46,8 +46,8 @@ void main() {
                 child: FButton(
                   label: const Text('Button'),
                   style: variant,
-                  prefix: FButtonIcon(icon: FAssets.icons.circlePlay),
-                  suffix: FButtonIcon(icon: FAssets.icons.circleStop),
+                  prefix: FIcon(FAssets.icons.circlePlay),
+                  suffix: FIcon(FAssets.icons.circleStop),
                   onPress: () {},
                 ),
               ),
@@ -77,8 +77,8 @@ void main() {
                 child: FButton(
                   label: const Text('Button'),
                   style: variant,
-                  prefix: FButtonIcon(icon: FAssets.icons.circlePlay),
-                  suffix: FButtonIcon(icon: FAssets.icons.circleStop),
+                  prefix: FIcon(FAssets.icons.circlePlay),
+                  suffix: FIcon(FAssets.icons.circleStop),
                   onPress: () {},
                 ),
               ),
@@ -108,8 +108,8 @@ void main() {
                 child: FButton(
                   label: const Text('Button'),
                   style: variant,
-                  prefix: FButtonIcon(icon: FAssets.icons.circlePlay),
-                  suffix: FButtonIcon(icon: FAssets.icons.circleStop),
+                  prefix: FIcon(FAssets.icons.circlePlay),
+                  suffix: FIcon(FAssets.icons.circleStop),
                   onPress: null,
                 ),
               ),
@@ -209,9 +209,7 @@ void main() {
                 child: FButton.icon(
                   onPress: () {},
                   style: variant,
-                  child: FButtonIcon(
-                    icon: FAssets.icons.chevronRight,
-                  ),
+                  child: FIcon(FAssets.icons.chevronRight),
                 ),
               ),
             ),
@@ -234,9 +232,7 @@ void main() {
                 child: FButton.icon(
                   onPress: null,
                   style: variant,
-                  child: FButtonIcon(
-                    icon: FAssets.icons.chevronRight,
-                  ),
+                  child: FIcon(FAssets.icons.chevronRight),
                 ),
               ),
             ),